+++ apps/application/src/components/Cases/CaseDocuments.tsx
@@ -12,10 +12,6 @@ import {
   CheckCircle,
   Clock,
   Loader2,
-  Folder,
-  FolderOpen,
-  ChevronRight,
-  Plus,
 } from "lucide-react";
 import {
   Tooltip,
@@ -23,12 +19,11 @@ import {
   TooltipProvider,
   TooltipTrigger,
 } from "../ui/tooltip";
-import { useEffect, useRef, useState } from "react";
+import { useState } from "react";
 import { IfCan } from "@/components/Permissions";
 import { PERMISSIONS } from "@/permissions/types";
 import { Id } from "convex/_generated/dataModel";
-import type { Folder as FolderType } from "../../../types/folders";
-import { Input } from "../ui/input";
+import { FolderTree } from "./FolderTree";
 
 interface CaseDocumentsProps {
   basePath: string;
@@ -43,23 +38,7 @@ export function CaseDocuments({ basePath }: CaseDocumentsProps) {
   const [currentFolderId, setCurrentFolderId] = useState<
     Id<"folders"> | undefined
   >(undefined);
-  const [isCreatingFolder, setIsCreatingFolder] = useState(false);
-  const [newFolderName, setNewFolderName] = useState("");
-  const newFolderInputRef = useRef<HTMLInputElement | null>(null);
-
-  // Fetch folders for the current case at current level
-  const folders = useQuery(
-    api.functions.folders.getFoldersForCase,
-    currentCase
-      ? { caseId: currentCase._id, parentFolderId: currentFolderId }
-      : "skip",
-  ) as FolderType[] | undefined;
-
-  // Fetch breadcrumb path when inside a folder
-  const breadcrumb = useQuery(
-    api.functions.folders.getFolderPath,
-    currentFolderId ? { folderId: currentFolderId } : "skip",
-  ) as FolderType[] | undefined;
+  // FolderTree handles fetching folders and breadcrumb
 
   // Fetch documents filtered by folder (root = no folderId)
   const documents = useQuery(
@@ -70,7 +49,6 @@ export function CaseDocuments({ basePath }: CaseDocumentsProps) {
   );
 
   const deleteDocument = useMutation(api.functions.documents.deleteDocument);
-  const createFolder = useMutation(api.functions.folders.createFolder);
 
   const handleDeleteDocument = async (documentId: Id<"documents">) => {
     setDeletingDocumentId(documentId);
@@ -81,30 +59,7 @@ export function CaseDocuments({ basePath }: CaseDocumentsProps) {
     }
   };
 
-  // Autofocus when opening the create input
-  useEffect(() => {
-    if (isCreatingFolder) {
-      const t = setTimeout(() => newFolderInputRef.current?.focus(), 0);
-      return () => clearTimeout(t);
-    }
-  }, [isCreatingFolder]);
-
-  const submitCreateFolder = async () => {
-    if (!currentCase) return;
-    const name = (newFolderName || "").trim() || "Nueva Carpeta";
-    try {
-      await createFolder({
-        name,
-        caseId: currentCase._id,
-        parentFolderId: currentFolderId,
-      } as any);
-      setNewFolderName("");
-      setIsCreatingFolder(false);
-    } catch (err) {
-      console.error("Error creating folder:", err);
-      alert(err instanceof Error ? err.message : "No se pudo crear la carpeta");
-    }
-  };
+  // Folder selection handled by FolderTree
 
   const getDocumentTypeColor = (documentType: string) => {
     switch (documentType) {
@@ -198,93 +153,13 @@ export function CaseDocuments({ basePath }: CaseDocumentsProps) {
 
   return (
     <div className="flex flex-col gap-2 pl-2 text-[12px] pt-1 overflow-y-auto max-h-32">
-      {/* Breadcrumb */}
-      <div className="flex items-center justify-between gap-1 text-muted-foreground">
-        <div className="flex items-center gap-1">
-          <button
-            className={`flex items-center gap-1 hover:text-foreground ${
-              !currentFolderId ? "text-foreground font-medium" : ""
-            }`}
-            onClick={() => setCurrentFolderId(undefined)}
-          >
-            <FolderOpen size={14} /> Raíz
-          </button>
-          {breadcrumb && breadcrumb.length > 0 && (
-            <>
-              {breadcrumb.map((f, idx) => (
-                <span key={f._id} className="flex items-center gap-1">
-                  <ChevronRight size={12} />
-                  <button
-                    className={`hover:text-foreground ${
-                      idx === breadcrumb.length - 1
-                        ? "text-foreground font-medium"
-                        : "text-muted-foreground"
-                    }`}
-                    onClick={() => setCurrentFolderId(f._id)}
-                  >
-                    {f.name}
-                  </button>
-                </span>
-              ))}
-            </>
-          )}
-        </div>
-        <IfCan permission={PERMISSIONS.DOC_WRITE} fallback={null}>
-          {!isCreatingFolder && (
-            <Button
-              variant="ghost"
-              size="sm"
-              className="h-6 px-2 py-0 hover:bg-gray-100"
-              onClick={() => setIsCreatingFolder(true)}
-            >
-              <Plus size={14} />
-            </Button>
-          )}
-        </IfCan>
-      </div>
-
-      {/* Inline new folder input */}
-      {isCreatingFolder && (
-        <div className="flex items-center gap-2 p-1">
-          <Input
-            ref={newFolderInputRef}
-            placeholder="Nombre de la carpeta"
-            value={newFolderName}
-            onChange={(e) => setNewFolderName(e.target.value)}
-            onKeyDown={(e) => {
-              if (e.key === "Enter") {
-                submitCreateFolder();
-              } else if (e.key === "Escape") {
-                setIsCreatingFolder(false);
-                setNewFolderName("");
-              }
-            }}
-            className="h-4 text-xs placeholder:text-xs"
-          />
-        </div>
-      )}
-
-      {/* Folders List */}
-      {folders && folders.length > 0 && (
-        <div className="flex flex-col gap-1">
-          {folders.map((folder) => (
-            <button
-              key={folder._id}
-              onClick={() => setCurrentFolderId(folder._id)}
-              className="flex items-center justify-between p-2 rounded hover:bg-gray-50"
-            >
-              <span className="flex items-center gap-2 min-w-0">
-                <Folder size={16} className="text-black flex-shrink-0" />
-                <span className="truncate">{folder.name}</span>
-              </span>
-              {folder.description && (
-                <span className="text-xs text-muted-foreground truncate max-w-[40%]">
-                  {folder.description}
-                </span>
-              )}
-            </button>
-          ))}
-        </div>
+      {/* Folder tree */}
+      {currentCase && (
+        <FolderTree
+          caseId={currentCase._id}
+          currentFolderId={currentFolderId}
+          onFolderChange={setCurrentFolderId}
+        />
       )}
 
       {/* Documents List */}
